requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docx-mailmerge==0.5.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
sh==1.14.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
whitenoise==5.2.0
